requirements.txt
backcall==0.2.0
blinker==1.4
click==8.1.3
decorator==5.1.1
Flask==1.1.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
Pygments==2.12.0
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.36
traitlets==5.2.1.post0
typing-extensions==4.2.0
wcwidth==0.2.5
Werkzeug==2.0.2
zipp==3.8.0
